Fix new warnings in Rust 1.88
We need this now because we'll soon be building with nightly so that the
C API can target x86_64-unknown-linux-none instead of
x86_64-unknown-none (for proper hard-float support).

Most of the changes here were just format string inlining, but 1.88 also
made `ControlFlow` a `must_use` type, so just explicitly discard
unneeded instances for now.