Build C API tests for `x86_64-unknown-linux-none`
This enables the built code to use hardware floating point, which should
both improve performance and enable us to pass floats across the C/Rust
ABI boundary in the future.

The `x86_64-unknown-linux-none` target is currently Tier 3 and doesn't
ship with a pre-built standard library, so we need to build with a
nightly toolchain for `-Z build-std` cargo support.