Display `fconst` values as hexfloats
This format is precise and isn't prone to any ambiguity/rounding when
parsing. It also means we no longer depend on floating-point display
being correct for exact roundtrips.

The only remaining problem is that we still don't differentiate between
distinct NaN bit patterns, but let's just ignore that for now.

Note that we currently depend on the git version of `hexfloat2` because
the most-recently published version predates the crate's `no_std`
support.