Add x64 codegen support for 64-bit-unsigned-to-float conversion
This is a bit more involved than other lowerings because it needs to
divide by 2 and then double after the conversion if bit 63 is set.

To get that working, create a new pseudo-operation and expand it into
the necessary branches during emission.

Note that this operation would benefit from tying one of its temporary
outputs to the input, but only if possible. We'll get around to that
once regalloc supports "soft ties".