Fix isel lowering of single-live-predecessor phis
Phi lowering incorrectly assumed that blocks with a single CFG
predecessor always choose phi input 0, but this is not correct because
the CFG does not contain dead control edges from the valgraph.

Fix this by choosing the correct input based on the valgraph predecessor
for single-CFG-predecessor blocks.

This was caught by one of TDN's unit tests. I'm not really sure whether
the dead block was intentional, but I sure am glad it was there!