Introduce regalloc "soft tie" constraint
The goal is for soft ties to act as hints that reusing a register
would improve codegen without forcing copies to be inserted when they
can't be satisfied. Unlike the existing ("hard") ties supported by the
allocator, multiple uses can be soft-tied to the same definition.

The motivating use case for soft ties is addition on x86, which we want
to be able to commute or convert into a `lea` based on surrounding
register constraints.

The allocator doesn't yet treat soft ties specially, but at least the
API exists now.