Introduce "mandatory" uncoalesced copy hints
In preparation for soft tie handling, track whether each copy hint is
"mandatory". A mandatory hint is one that will cause a copy to be
inserted later if broken.

The only effect mandatory hints have is on whether conflicts on
rematerializable fragments are reported eagerly or not: if a fragment
has several inputs tied to the same output, we don't want the inevitable
conflict there to force premature rematerialization. That logic might
need to be revisited in the future if it has too many adverse side
effects.

For now, all hints are mandatory because they all come from ties/block
params.