Tweak regalloc double-split handling
The double-split logic assumed that the range obtained by splitting at
the first split point would always be splittable at the second, but this
is incorrect for ranges with holes around the second split point.

Make sure double split handling always checks whether the range is
splittable at the second split point before going through with it, and
simplify some code in the double-split computation now that that is the
case.