Store function metadata separately from bodies
This is a large-scale reorganization of how functions are stored in a
module, so it had quite a bit of fallout.

The goal is to enable a function body to be (mutably) borrowed while
retaining access to the module-global metadata map, which we currently
need in a few places (frontend, isel, verification, display) and will
wind up needing in a few more.

Most of the refactor was mechanical, so there are still a large number
of APIs that need to be cleaned up.