Make `ModuleMetadata` fields private
This helps enforce the invariant that a module's metadata should only be
mutated through the owning module itself.

Note that the fields inside `ModuleMetadata` were made private rather
than the whole `metadata` field in `Module` because we need to be able
to borrow `metadata` separately from `functions`. It does make for a bit
of an awkward API, but it shouldn't be too bad.