Do away with `FunctionData`
After internal function metadata was moved into a separate map,
`FunctionData` became an awkward structure holding just the body and the
node cache.

Most clients don't care about the node cache, and it was awkward to have
to work harder to reach the interesting part (the body), so just move
the node caches to their own side table and get rid of `FunctionData`
completely.