Refactor internal reducer state again
The goal this time is to remove borrows in `ReducerState` itself so it
can be held onto for longer (i.e., across reduction cycles) later.

Doing that alone would have caused `ReducerBuilder` to have the same
fields as `ReduceContext` with no tangible benefit, so the `Builer`
implementation has also been folded directly into `ReduceContext`. This
actually cleans up some boilerplate in the canonicalizer, which is nice.