Introduce new `RegWidth` and `RegClass` types
A "class" now ties a bank (what used to be referred to as the "class")
and a "width", which is a marker indicating what portion of the register
is used. The goal is to eventually be able to emit narrower
spills/copies where possible; that should be especially beneficial
around SSE registers.