Introduce dedicated `Spill` parallel copy phase
Commit aeba278 ("Rework regalloc assignment copy phases") introduced a
dedicated `Reload` copy phase so that spills from the previous
instruction into reloaded spill slot are guaranteed to be performed
before the reload is.

While that dedicated phase is indeed necessary, the analysis there
forgot to take copies to/from spill slots into account. The core
allocator always treats copies as happening in parallel "between"
instructions, while spills and reloads always "belong" to the
instructions using them.

In particular, if adjacent instructions A and B respectively spill into
and reload from a particular spill slot, and there are copies to or from
that same spill slot between the two, the inserted copies must be
ordered as:

    spill
    copies
    reload

for the correct values to propagate. The way the code was written
previously, the spill and the subsequent copies targeting the slot would
all be performed in parallel, leading to stale values being read in
certain cases.

This was caught by one of TDN's unoptimized test cases with the
following verifier error:

    expected i482 use 0 to be of %4594, found {%4578, %966, %1031, %1097, %1120}

It has now been added as a regression test.