Stop checking reservations during register scavenging
More careful reasoning about global allocator invariants shows that a
register will never actually need to be scavenged at points where a
reservation is live, so we don't actually need to bother with checking
the reservations in the first place. See the new assertions and
comments in the code for details.

The reason we care about this isn't some perceived optimization
opportunity (although the change likely can't hurt); rather, introducing
support for overlapping register classes with different widths (such as
32-bit vs 64-bit GPRs on x64) will require scavenging to report more
detailed information in certain cases. That information would be a bit
unwieldy to store/retrieve for reservations, so make sure we won't even
need to try in the future.