Change parallel copy register scavenging interface
Now, rather than just providing temporary registers when the parallel
copy resolver asks for them, the scavenger is responsible for tracking
the allocation status of returned temporaries itself. This was pretty
easy to wire up because both implementations were already doing that for
other reasons.

Beyond simplifying the code somewhat, this change improved codegen in
cases involving disjoint stack-to-stack copies and cycles in the same
resolution.