Stop returning bank information with parallel copy chunks
We no longer need to know the bank outside because the resolver now
passes it wherever necessary.

Note that, despite this change, parallel copies are still resolved
independently on a per-bank basis. The reasoning is that it's probably
still better to do multiple small invocations than a single large one,
though we may want to examine that more thoroughly in the future.