Rename x64 "full" register classes
This is the first step on the way to actually supporting multiple
register widths in the same bank.

Note that the XMM register class has been renamed to `xmm64` and not
`xmm128`, because we don't actually need the full width of the registers
anywhere right now.

No functional change for now, but a whole bunch of test output has
changed.