Create x64 `gpr32` register class
It is now used for all 32-bit integer values, but doesn't yet affect
codegen.

The one thing I'm still a bit uneasy about is the fact that the verifier
doesn't enforce anything about the widths of emitted copies; this was
just too hard to check without false positives given that we allow block
parameter copies to change register width.