Overhaul representation of optimizer state
The goal here is to eventually have a long-lived "state" object that
exists throughout the entirety of the optimization pipeline, with all
optimization passes referring to it.

Canonicalization is a bit special in that we want to be able to
incrementally canonicalize nodes created/updated by other passes
throughout the entire pipeline, so some canonicalization-specific state
needs to be maintained as well.

To that effect:

- Get rid of the generic "reducer" API (it only really makes sense for
  canonicalization, at least as far as I can tell for the moment).

- Rename `ReduceState` to `FunctionState`; it will eventually be the
  object that lives throughout the optimization pipeline.

- Rename `ReduceContext` to `EditContext`. This transient object can be
  created whenever a pass wants to edit the content of a specific
  function.